+++ job1/src/App.js
@@ -6,7 +6,7 @@ import Laws from "./View/Laws";
 import Board from "./View/Board";
 import News from "./View/News";
 import Footer from "./Main/Footer";
-import Login from "./View/Login";
+import Login from "./View/Login/Login";
 import MyPage from "./View/MyPage";
 import Write from "./View/Write";
 import Tag from "./Tags/Tag";
