--- job1/src/View/Login.js
@@ -1,11 +1,0 @@
-import SocialKakao from '../Component/LoginComponent'
-
-function Login(){
-    return(
-        <div>
-            <SocialKakao/>
-        </div>
-    )
-}
-
-export default Login;
